Fix __str__ and __unicode__ methods
The behavior of the `__str__` method between Python 2 and 3 is
confusing! In Python 2 it returns a binary type value and in Python 3 it
returns a unicode type value. In Python 2, the `__unicode__` method
returns a unicode type value and in Python 3 this method never gets
called.

The previous implementation sort of reversed the correct behavior.

Since we're already using the `six` package, the
`@six.python_2_unicode_compatible` decorator seems like a good way to
make both the `__str__` and `__unicode__` methods return the correct
values. In particular, if we ever wanted to drop Python 2 support, we
would just have to drop the decorator.

Addresses #121
https://github.com/datadesk/python-documentcloud/issues/121